// backend/src/generated/models/AuditLog.ts
export type AuditLogWhereInput = {
  AND?: Prisma.AuditLogWhereInput | Prisma.AuditLogWhereInput[]
  OR?: Prisma.AuditLogWhereInput[]
  NOT?: Prisma.AuditLogWhereInput | Prisma.AuditLogWhereInput[]
  id?: Prisma.StringFilter<"AuditLog"> | string
  timestamp?: Prisma.DateTimeFilter<"AuditLog"> | Date | string
  severity?: Prisma.EnumAuditLogSeverityFilter<"AuditLog"> | $Enums.AuditLogSeverity
  event?: Prisma.StringFilter<"AuditLog"> | string
  status?: Prisma.EnumAuditLogStatusFilter<"AuditLog"> | $Enums.AuditLogStatus
  actorId?: Prisma.StringFilter<"AuditLog"> | string
  actorName?: Prisma.StringFilter<"AuditLog"> | string
  actorType?: Prisma.EnumActorTypeFilter<"AuditLog"> | $Enums.ActorType
  actorIp?: Prisma.StringFilter<"AuditLog"> | string
  actorLocation?: Prisma.StringNullableFilter<"AuditLog"> | string | null
  targetId?: Prisma.StringNullableFilter<"AuditLog"> | string | null
  targetName?: Prisma.StringNullableFilter<"AuditLog"> | string | null
  targetType?: Prisma.StringNullableFilter<"AuditLog"> | string | null
  metadata?: Prisma.JsonNullableFilter<"AuditLog">
  userId?: Prisma.StringNullableFilter<"AuditLog"> | string | null
  deviceId?: Prisma.StringNullableFilter<"AuditLog"> | string | null
  apiKeyId?: Prisma.StringNullableFilter<"AuditLog"> | string | null
  user?: Prisma.XOR<Prisma.UserNullableScalarRelationFilter, Prisma.UserWhereInput> | null
  device?: Prisma.XOR<Prisma.DeviceNullableScalarRelationFilter, Prisma.DeviceWhereInput> | null
  apiKey?: Prisma.XOR<Prisma.ApiKeyNullableScalarRelationFilter, Prisma.ApiKeyWhereInput> | null
}